requirements.txt
python-dotenv
google-generativeai
streamlit
PYPDF2